requirements.txt
python-telegram-bot>=20.0
fastapi
uvicorn
python-multipart
python-dotenv
